swiftmotion_server/requirements.txt
celery==4.0.2
Django==1.10.5
django-bootstrap3==8.2.1
djangorestframework==3.5.4
psycopg2==2.6.2
redis==2.10.5
